perf: enable faster Ace bootstrap with targeted exports
