+++ index.ts
@@ -15,7 +15,7 @@ import { errors as httpServerErrors } from '@adonisjs/http-server'
 
 export { stubsRoot } from './stubs/main.ts'
 export { inject } from './modules/container.ts'
-export { Ignitor } from './src/ignitor/main.ts'
+export { Ignitor, prettyPrintError } from './src/ignitor/main.ts'
 export { configProvider } from './src/config_provider.ts'
 export { indexEntities } from './src/assembler_hooks/index_entities.ts'
 
@@ -32,23 +32,4 @@ export const errors: typeof encryptionErrors &
   ...appErrors,
   ...aceErrors,
   ...envErrors,
-}
-
-/**
- * Pretty prints an error with colorful output using
- * Youch terminal
- */
-export async function prettyPrintError(error: any) {
-  if (error && typeof error === 'object' && error.code === 'E_DUMP_DIE_EXCEPTION') {
-    console.error(error)
-    return
-  }
-
-  try {
-    const { Youch } = await import('youch')
-    const youch = new Youch()
-    console.error(await youch.toANSI(error))
-  } catch {
-    console.error(error)
-  }
 }
