+++ package.json
@@ -27,6 +27,8 @@
     "./providers/*": "./build/providers/*.js",
     "./helpers": "./build/src/helpers/main.js",
     "./helpers/*": "./build/src/helpers/*.js",
+    "./ignitor": "./build/src/ignitor/main.js",
+    "./generators": "./build/src/assembler_hooks/index_entities.js",
     "./ace": "./build/modules/ace/main.js",
     "./ace/codemods": "./build/modules/ace/codemods.js",
     "./bodyparser": "./build/modules/bodyparser/main.js",
@@ -235,6 +237,8 @@
       "providers/**/*.ts",
       "src/helpers/main.ts",
       "src/helpers/**/*.ts",
+      "src/ignitor/main.ts",
+      "src/assembler_hooks/index_entities.ts",
       "modules/ace/main.ts",
       "modules/ace/codemods.ts",
       "modules/bodyparser/main.ts",
